requirements.txt
fastapi
uvicorn
python-dotenv
openai
pypdf
fpdf
python-multipart